requirements.txt
asgiref==3.5.2
channels==3.0.5
dj_rest_auth==2.2.5
Django==3.2 #4.1.5
django_environ==0.9.0
django_filter==22.1
django_heroku==0.3.1
environ==1.0
Pillow==9.2.0
requests==2.28.1
whitenoise==6.3.0
djangorestframework==3.14.0
pandas==1.5.3
openpyxl==3.0.9
python-dotenv==0.19.1
django-crispy-forms==1.12.0
django_cleanup==5.1.0
django-ckeditor==6.1.0
